+++ packages/terra-popup/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Popup Content will clone prop to indicate missing close button
 
 4.9.0 - (July 26, 2018)
 ------------------
@@ -12,7 +14,6 @@ Unreleased
 4.8.0 - (July 25, 2018)
 ------------------
 ### Changed
-* Popup Content will clone prop to indicate missing close button
 * Add .module extension to SCSS file(s)
 
 4.7.0 - (July 10, 2018)